[CHORE] #1 - move base classes to other package
