+++ app/src/main/java/org/mydaily/ui/base/BaseFragment.kt
@@ -1,4 +1,4 @@
-package org.mydaily.ui.view.base
+package org.mydaily.ui.base
 
 import android.os.Bundle
 import android.view.LayoutInflater
