+++ app/src/main/java/org/mydaily/ui/view/mypage/MyPageFragment.kt
@@ -3,7 +3,7 @@ package org.mydaily.ui.view.mypage
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import org.mydaily.R
 import org.mydaily.databinding.FragmentMyPageBinding
-import org.mydaily.ui.view.base.BaseFragment
+import org.mydaily.ui.base.BaseFragment
 import org.mydaily.ui.viewmodel.MyPageViewModel
 
 class MyPageFragment : BaseFragment<FragmentMyPageBinding, MyPageViewModel>() {
