+++ app/src/main/java/org/mydaily/ui/view/remind/RemindFragment.kt
@@ -3,7 +3,7 @@ package org.mydaily.ui.view.remind
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import org.mydaily.R
 import org.mydaily.databinding.FragmentRemindBinding
-import org.mydaily.ui.view.base.BaseFragment
+import org.mydaily.ui.base.BaseFragment
 import org.mydaily.ui.viewmodel.RemindViewModel
