+++ app/src/main/java/org/mydaily/ui/viewmodel/DailyViewModel.kt
@@ -1,6 +1,6 @@
 package org.mydaily.ui.viewmodel
 
-import org.mydaily.ui.view.base.BaseViewModel
+import org.mydaily.ui.base.BaseViewModel
 
 class DailyViewModel: BaseViewModel() {
 }
